Better fix for marking fs err in superblock if err set in journal superblock.
The previous fix of flushing out the superblock after setting the
EXT2_ERROR_FS flag was erroneous because running the journal could have
updated the superblock (but not the in-memory copy the code referenced).
So instead, check the journal error conditions after running the journal
and freeing and re-opening the filesystem.  Also, make sure to flush
the error state of the superblock to disk before clearing the error state
of the journal.

Change-Id: Ia5cde1a367d86992c9bd3bca652a1289935d1380